+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Add the `Password` dialog box.
+
 # v0.1.0 (2019-01-08)
 - Initial release with the `Input`, `Message` and `Question` dialog boxes and
   the `Dialog` backend.